requirements.txt
numpy
minio
google-cloud-bigquery
python-dotenv
